TMS-134: Extract generic uniqueId behaviour into a base class
