+++ src/zeit/cms/repository/file.py
@@ -1,19 +1,18 @@
-
 from zeit.cms.i18n import MessageFactory as _
 import ZODB.blob
 import persistent
 import zeit.cms.interfaces
 import zeit.cms.repository.interfaces
+import zeit.cms.repository.repository
 import zeit.cms.type
 import zeit.cms.workingcopy.interfaces
 import zeit.connector.interfaces
-import zope.app.container.contained
 import zope.component
 import zope.interface
 import zope.security.proxy
 
 
-class RepositoryFile(zope.app.container.contained.Contained):
+class RepositoryFile(zeit.cms.repository.repository.ContentBase):
     """A file in the repository."""
 
     zope.interface.implements(zeit.cms.repository.interfaces.IFile,
